Tested DShot using timers + DMA on one channel. Need to update it for 4 channels

+++ firmware/no_fly_zone/Core/Inc/dshot.h
@@ -53,6 +53,6 @@ typedef struct {
 } DShotParams;
 
 void dshot_init();
-void dshot_encode(uint16_t throttle, uint8_t tel_req, uint16_t * pkt);
+void dshot_encode(uint16_t throttle, uint8_t tel_req, uint32_t * pkt);
 
 #endif
